refactor(LoggedInMember): 로그인한 멤버를 dto 패키지로 이동
- class 대신 record 사용
